refactor : move request/response class to dto
