+++ src/main/kotlin/com/group4/ticketingservice/dto/SignDTO.kt
@@ -1,0 +1,12 @@
+package com.group4.ticketingservice.dto
+
+data class SignUpRequest(
+        val email: String,
+        val name: String,
+        val password: String)
+
+data class SignInRequest(
+        val email: String,
+        val password: String
+
+)
